Fix Screen Handling on Sonoma
On Sonoma, getting the right screen coordiantes was broken (Issue #517).
Somhow the old CG code stopped working, so the fix was to use more modern
APIs.

Getting the screen height correctly on macs with a notch remnained tricky.
Uebersicht has a weird use case where you want 'full screen' for the most part
except for drawing under the top menu bar.

Remaining backwards compatible will be tricky, plus this issue is only present
on macOS 14, so this bumps the minimum requirements to macOS 12.